+++ packages/plugin-gallery/web/src/defaults.ts
@@ -1,7 +1,22 @@
 import { ImageComponentData, VideoComponentData } from 'wix-rich-content-common';
-import { GALLERY_LAYOUTS } from '../lib/layout-data-provider';
 
 /* eslint-disable camelcase */
+export const GALLERY_LAYOUTS = Object.freeze({
+  EMPTY: -1,
+  COLLAGE: 0,
+  MASONRY: 1,
+  GRID: 2,
+  THUMBNAIL: 3,
+  SLIDER: 4,
+  SLIDESHOW: 5,
+  PANORAMA: 6,
+  COLUMN: 7,
+  MAGIC: 8,
+  FULLSIZE: 9,
+  BRICKS: 10,
+  MIX: 11,
+  ALTERNATE: 12,
+});
 
 export const GALLERY_ITEMS_TYPES = Object.freeze({
   IMAGE: 'image',
@@ -81,23 +96,21 @@ export const createImageItem = (
     },
     itemId,
     url: preloadImage ? img.src : img.file_name,
-    tempData: preloadImage,
+    tempData: true,
   };
 };
 
 export const createVideoItem = (video: VideoComponentData, itemId: string) => {
-  const {
-    thumbnail: { pathname: poster, width, height },
-  } = video;
   return {
     metadata: {
       type: 'video',
-      height: video.height || height,
-      width: video.width || width,
-      poster,
+      height: video.thumbnail.height,
+      width: video.thumbnail.width,
+      poster: video.thumbnail.pathname,
     },
     itemId,
     url: video.pathname,
+    tempData: true,
   };
 };