+++ packages/ricos-content/web/src/RicosContentAPI/extract.ts
@@ -1,0 +1,5 @@
+// import * as T from 'fp-ts/lib/Tree';
+// import { fromTraversable } from 'monocle-ts';
+import { RichContent } from 'ricos-schema';
+
+export const extract = (content: RichContent) => {};